Adds neo4j community under GPLv3

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/CountsRecordState.java
@@ -44,9 +44,9 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.neo4j.kernel.impl.storageengine.impl.recordstorage.RecordState;
 import org.neo4j.kernel.impl.store.counts.keys.CountsKey;
 import org.neo4j.kernel.impl.transaction.command.Command;
-import org.neo4j.kernel.impl.transaction.state.RecordState;
 import org.neo4j.register.Register.DoubleLongRegister;
 import org.neo4j.register.Registers;
 import org.neo4j.storageengine.api.StorageCommand;
@@ -73,7 +73,7 @@ public class CountsRecordState implements CountsAccessor, RecordState, CountsAcc
     }
 
     @Override
-    public void incrementNodeCount( int labelId, long delta )
+    public void incrementNodeCount( long labelId, long delta )
     {
         counts( nodeKey( labelId ) ).increment( 0L, delta );
     }
@@ -94,7 +94,7 @@ public class CountsRecordState implements CountsAccessor, RecordState, CountsAcc
     }
 
     @Override
-    public void incrementRelationshipCount( int startLabelId, int typeId, int endLabelId, long delta )
+    public void incrementRelationshipCount( long startLabelId, int typeId, long endLabelId, long delta )
     {
         if ( delta != 0 )
         {
