+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/counts/CountsUpdater.java
@@ -71,7 +71,7 @@ final class CountsUpdater implements CountsAccessor.Updater, CountsAccessor.Inde
      * For key format, see {@link KeyFormat#visitNodeCount(int, long)}
      */
     @Override
-    public void incrementNodeCount( int labelId, long delta )
+    public void incrementNodeCount( long labelId, long delta )
     {
         try
         {
@@ -93,7 +93,7 @@ final class CountsUpdater implements CountsAccessor.Updater, CountsAccessor.Inde
      * For key format, see {@link KeyFormat#visitRelationshipCount(int, int, int, long)}
      */
     @Override
-    public void incrementRelationshipCount( int startLabelId, int typeId, int endLabelId, long delta )
+    public void incrementRelationshipCount( long startLabelId, int typeId, long endLabelId, long delta )
     {
         try
         {
